# `fold_unrolled_divs` with float cast 