--- a/test/unit/test_uop_symbolic.py
+++ b/test/unit/test_uop_symbolic.py
@@ -423,6 +423,10 @@ class TestSymbolic(unittest.TestCase):
     self.helper_test_variable((gidx0+lidx2+lidx3)*4, 0, 80,
                               ("(((gidx0*4)+(lidx2*4))+(lidx3*4))","((lidx3*4)+((gidx0*4)+(lidx2*4)))"))
 
+  def test_float_mul_over_int_cast_distribute(self):
+    gidx0 = Variable("gidx0", 0, 7)
+    self.helper_test_variable((gidx0+10).cast(dtypes.float)*0.5, float("-inf"), float("inf"), "(((float)(gidx0)*0.5)+5.0)")
+
   @unittest.expectedFailure
   def test_variable_divmod(self):
     start_pos = Variable("start_pos", 0, 127)
@@ -486,6 +490,11 @@ class TestSymbolic(unittest.TestCase):
     gidx = Variable("gidx", 0, 2559)
     unrolled_div = (gidx+2561)//4+(gidx+2562)//4+(gidx+2560)//4+(gidx+2559)//4
     self.helper_test_variable(unrolled_div, 2559, 5118, "(gidx+2559)")
+
+  def test_arange_unrolled4_float_cast(self):
+    ridx0 = Variable("ridx0", 0, 63)
+    expr = ((ridx0+1)//4).cast(dtypes.float)*0.75+((ridx0+2)//4).cast(dtypes.float)*0.75+((ridx0+3)//4).cast(dtypes.float)*0.75+((ridx0//4)+1).cast(dtypes.float)*0.75+-30.95
+    self.helper_test_variable(expr, float("-inf"), float("inf"), "((0.75*(float)(ridx0))+-30.2)")
 
   def test_arange_unrolled4_mul(self):
     gidx = Variable("gidx", 0, 2559)