Ill make the symbolic flat rule a seperate pr as well