@S-Lykles 2025-05-15
Not really showing process replay but running `DEBUG=4 pytest -s test/test_ops.py::TestOps::test_arange` gives
master:
```
*** CUDA      13 r_17_4_17_4                               arg  1 mem  0.00 GB tm     37.12us/     0.83ms (     0.03 GFLOPS    0.0|0.0     GB/s) ['arange']
*** CPU       14 copy      272,     CPU <- CUDA            arg  2 mem  0.00 GB tm    176.98us/     1.01ms (     0.00 GFLOPS    0.0|0.0     GB/s)

testing                                       []   torch/tinygrad fp: 0.05 / 62.41 ms  bp: nan / nan ms r_5_8_10_4
 0: (5, 8, 1, 1)              float.ptr(40)        (8, 1, 0, 0)                   ShapeTracker(views=(View(shape=(5, 8, 1, 1), strides=(8, 1, 0, 0), offset=0, mask=None, contiguous=True),))
[Opt(op=OptOps.UNROLL, axis=0, arg=4), Opt(op=OptOps.LOCAL, axis=0, arg=8)]
#define INFINITY (__int_as_float(0x7f800000))
#define NAN (__int_as_float(0x7fffffff))
extern "C" __global__ void __launch_bounds__(8) r_5_8_10_4(float* data0) {
  int gidx0 = blockIdx.x; /* 5 */
  int lidx0 = threadIdx.x; /* 8 */
  int alu0 = (gidx0<<1);
  *(data0+(lidx0+(gidx0<<3))) = ((((float)((alu0+((lidx0+1)>>2))))*0.75f)+(((float)((alu0+((lidx0+2)>>2))))*0.75f)+(((float)((alu0+((lidx0+3)>>2))))*0.75f)+(((float)((alu0+(lidx0>>2)+1)))*0.75f)+-30.95f);
}
```

This branch
```
*** CUDA      13 r_17_4_17_4                               arg  1 mem  0.00 GB tm     31.36us/     0.89ms (     0.00 GFLOPS    0.0|0.0     GB/s) ['arange']
*** CPU       14 copy      272,     CPU <- CUDA            arg  2 mem  0.00 GB tm    107.91us/     1.00ms (     0.00 GFLOPS    0.0|0.0     GB/s)

testing                                       []   torch/tinygrad fp: 0.05 / 18.40 ms  bp: nan / nan ms r_5_8_10_4
 0: (5, 8, 1, 1)              float.ptr(40)        (8, 1, 0, 0)                   ShapeTracker(views=(View(shape=(5, 8, 1, 1), strides=(8, 1, 0, 0), offset=0, mask=None, contiguous=True),))
[Opt(op=OptOps.UNROLL, axis=0, arg=4), Opt(op=OptOps.LOCAL, axis=0, arg=8)]
#define INFINITY (__int_as_float(0x7f800000))
#define NAN (__int_as_float(0x7fffffff))
extern "C" __global__ void __launch_bounds__(8) r_5_8_10_4(float* data0) {
  int gidx0 = blockIdx.x; /* 5 */
  int lidx0 = threadIdx.x; /* 8 */
  int alu0 = (lidx0+(gidx0<<3));
  *(data0+alu0) = ((0.75f*((float)(alu0)))+-30.2f);
}
```