@S-Lykles 2025-05-22
```
'test_ops.py:2502 test_interpolate_linear_corners_aligned'
--- 

+++ 

@@ -3,9 +3,10 @@

 kernel void r_13_4_13_4n3(device float* data0, uint3 gid [[threadgroup_position_in_grid]], uint3 lid [[thread_position_in_threadgroup]]) {
   int gidx0 = gid.x; /* 13 */
   int lidx0 = lid.x; /* 4 */
-  float alu0 = (0.5490196078431373f*(((float)((gidx0+((lidx0+1)>>2))))+((float)((gidx0+((lidx0+2)>>2))))+((float)((gidx0+((lidx0+3)>>2))))+((float)((gidx0+1)))+-1.0f));
-  float alu1 = ((alu0<0.0f)?0.0f:alu0);
-  float alu2 = -alu1;
-  float alu3 = ((alu2<-28.0f)?-28.0f:alu2);
-  *(data0+(lidx0+(gidx0<<2))) = -alu3;
+  int alu0 = (lidx0+(gidx0<<2));
+  float alu1 = (0.5490196078431373f*(((float)((alu0+1)))+-1.0f));
+  float alu2 = ((alu1<0.0f)?0.0f:alu1);
+  float alu3 = -alu2;
+  float alu4 = ((alu3<-28.0f)?-28.0f:alu3);
+  *(data0+alu0) = -alu4;
 }
 ```